requirements.txt
folium==0.16.0
haversine==2.8.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
streamlit==1.34.0
streamlit_folium==0.20.0
